moved admin methods into donationswap class (added new admin ajax decorator)
